1、数据类型修改

+++ business/play_base/src/main/java/com/will/play/base/entity/PickDouyinEntity.kt
@@ -98,7 +98,7 @@ data class VideoInfo(
     val linkurl: String,
     val material_lib_type: String,
     val max_commission: String,
-    val max_commission_rate: Int,
+    val max_commission_rate: Float,
     val max_commission_rate_text: String,
     val mobile: String,
     val nick: String,
